+++ olive/model/handler/onnx.py
@@ -17,6 +17,7 @@ from olive.hardware.accelerator import AcceleratorLookup, Device
 from olive.model.config.registry import model_handler_registry
 from olive.model.handler.base import OliveModelHandler
 from olive.model.mixin import OnnxEpValidateMixin, OnnxGraphMixin
+from olive.model.utils.onnx_utils import get_onnx_file_path
 from olive.resource_path import OLIVE_RESOURCE_ANNOTATIONS
 
 logger = logging.getLogger(__name__)
@@ -118,66 +119,6 @@ class ONNXModelHandler(OliveModelHandler, OnnxEpValidateMixin, OnnxGraphMixin):
         # save io_config
         self.io_config = self.get_graph_io_config()
         return self.io_config
-
-
-def resolve_path(file_or_dir_path: str, model_filename: str = "model.onnx") -> str:
-    """Get the model full path.
-
-    The engine provides output paths to ONNX passes that do not contain .onnx extension
-    (these paths are generally locations in the cache). This function will convert such
-    paths to absolute file paths and also ensure the parent directories exist.
-    If the input path is already an ONNX file it is simply returned. Examples:
-
-    resolve_path("c:/foo/bar.onnx") -> c:/foo/bar.onnx
-
-    resolve_path("c:/foo/bar") -> c:/foo/bar/model.onnx
-    """
-    if not model_filename.endswith(".onnx"):
-        raise ValueError(f"ONNXModel's model name must end with '.onnx', got {model_filename}")
-
-    path = Path(file_or_dir_path)
-    if path.suffix != ".onnx":
-        path = path / model_filename
-        parent_dir = path.parent
-        if not parent_dir.exists():
-            parent_dir.mkdir(parents=True, exist_ok=True)
-    return str(path)
-
-
-def get_onnx_file_path(model_path: str, onnx_file_name: Optional[str] = None) -> str:
-    """Get the path to the ONNX model file.
-
-    If model_path is a file, it is returned as is. If model_path is a
-    directory, the onnx_file_name is appended to it and the resulting path is returned. If onnx_file_name is not
-    specified, it is inferred if there is only one .onnx file in the directory, else an error is raised.
-    """
-    assert Path(model_path).exists(), f"Model path {model_path} does not exist"
-
-    # if model_path is a file, return it as is
-    if Path(model_path).is_file():
-        return model_path
-
-    # if model_path is a directory, append onnx_file_name to it
-    if onnx_file_name:
-        onnx_file_path = Path(model_path) / onnx_file_name
-        assert onnx_file_path.exists(), f"ONNX model file {onnx_file_path} does not exist"
-        return str(onnx_file_path)
-
-    # try to infer onnx_file_name
-    logger.warning(
-        "model_path is a directory but onnx_file_name is not specified. Trying to infer it. It is recommended to"
-        " specify onnx_file_name explicitly."
-    )
-    onnx_file_names = list(Path(model_path).glob("*.onnx"))
-    if len(onnx_file_names) == 1:
-        return str(onnx_file_names[0])
-    elif len(onnx_file_names) > 1:
-        raise ValueError(
-            f"Multiple .onnx model files found in the model folder {model_path}. Please specify one using the"
-            " onnx_file_name argument."
-        )
-    else:
-        raise ValueError(f"No .onnx file found in the model folder {model_path}.")
 
 
 @model_handler_registry("DistributedOnnxModel")
