+++ olive/model/utils/__init__.py
@@ -1,0 +1,13 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+# --------------------------------------------------------------------------
+from olive.model.utils.hf_mappings import HIDDEN_SIZE_NAMES, MODEL_TYPE_MAPPING, NUM_HEADS_NAMES
+from olive.model.utils.onnx_utils import resolve_path
+
+__all__ = [
+    "HIDDEN_SIZE_NAMES",
+    "MODEL_TYPE_MAPPING",
+    "NUM_HEADS_NAMES",
+    "resolve_path",
+]
