+++ olive/passes/onnx/append_pre_post_processing_ops.py
@@ -12,7 +12,8 @@ from packaging import version
 from olive.common.config_utils import ConfigBase
 from olive.common.pydantic_v1 import Field, validator
 from olive.hardware.accelerator import AcceleratorSpec
-from olive.model import ONNXModelHandler, resolve_path
+from olive.model import ONNXModelHandler
+from olive.model.utils import resolve_path
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_olive_model
 from olive.passes.onnx.pipeline import TENSOR_TYPE_MAP
