+++ olive/passes/onnx/conversion.py
@@ -24,9 +24,9 @@ from olive.model import (
     HfFromPretrainedArgs,
     ONNXModelHandler,
     PyTorchModelHandler,
-    resolve_path,
 )
 from olive.model.config import IoConfig
+from olive.model.utils import resolve_path
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_olive_model
 from olive.passes.pass_config import PassConfigParam
