+++ olive/passes/onnx/float16_conversion.py
@@ -8,7 +8,8 @@ from typing import Any, Dict, List
 import onnx
 
 from olive.hardware.accelerator import AcceleratorSpec
-from olive.model import ONNXModelHandler, resolve_path
+from olive.model import ONNXModelHandler
+from olive.model.utils import resolve_path
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_olive_model
 from olive.passes.pass_config import PassConfigParam
