+++ olive/passes/onnx/insert_beam_search.py
@@ -9,7 +9,8 @@ from onnx import ModelProto, TensorProto, helper
 from packaging import version
 
 from olive.hardware.accelerator import AcceleratorSpec, Device
-from olive.model import CompositeModelHandler, OliveModelHandler, ONNXModelHandler, resolve_path
+from olive.model import CompositeModelHandler, OliveModelHandler, ONNXModelHandler
+from olive.model.utils import resolve_path
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_olive_model
 from olive.passes.pass_config import PassConfigParam
