+++ olive/passes/onnx/transformer_optimization.py
@@ -10,8 +10,8 @@ from typing import TYPE_CHECKING, Any, Dict, List, Union
 import onnx
 
 from olive.hardware.accelerator import AcceleratorSpec, Device
-from olive.model import ONNXModelHandler, resolve_path
-from olive.model.hf_mappings import HIDDEN_SIZE_NAMES, MODEL_TYPE_MAPPING, NUM_HEADS_NAMES
+from olive.model import ONNXModelHandler
+from olive.model.utils import HIDDEN_SIZE_NAMES, MODEL_TYPE_MAPPING, NUM_HEADS_NAMES, resolve_path
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_olive_model
 from olive.passes.pass_config import PassConfigParam
