+++ olive/passes/onnx/vitis_ai_quantization.py
@@ -13,7 +13,8 @@ import onnx
 from olive.cache import get_local_path_from_root
 from olive.common.utils import hash_string
 from olive.hardware import AcceleratorSpec
-from olive.model import ONNXModelHandler, resolve_path
+from olive.model import ONNXModelHandler
+from olive.model.utils import resolve_path
 from olive.passes import Pass
 from olive.passes.onnx.common import get_external_data_config, model_proto_to_file, model_proto_to_olive_model
 from olive.passes.pass_config import ParamCategory, PassConfigParam
