+++ olive/passes/pytorch/sparsegpt_utils.py
@@ -13,7 +13,7 @@ import torch
 import transformers
 
 from olive.common.utils import get_attr
-from olive.model.hf_mappings import MODELS_TO_EMBEDDINGS_MAPPING, MODELS_TO_LAYERS_MAPPING
+from olive.model.utils.hf_mappings import MODELS_TO_EMBEDDINGS_MAPPING, MODELS_TO_LAYERS_MAPPING
 
 logger = logging.getLogger(__name__)
